[wip] Monkey-patch `Hanami::Action.call` method
Instead of instrumenting Rack requests, instrument calls to the
`call` method in `Hanami::Action` (which receives a Rack request
as an argument) so that we can use the action's class name as the
action name.
